+++ config/routes.rb
@@ -56,6 +56,7 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
+      get   'health', to: 'health#index'
       patch 'settings', to: 'settings#update'
       get   'settings', to: 'settings#index'